Refactor_202412.08.13.07_MatrixViewコンポーネントのリファクタリング
